Revert "FunkyChem (#268)" (#978)
CONTROVERSAL PR EXPLANATION:
A lot of this boils down to how switching to Den is... awful, for
skilled chemists. Some may not be as effected, but people like me, we go
from server to server and find it incredibly difficult to adjust for any
meaningful period of time because of that. It's rough.

There really isn't a need to have this anymore, given how difficult it
is to transfer and how it doesn't really provide much that we can't do
ourselves. Hell, at worst, I just add to WizChem to make it better.

It's way too easy, heals things way too fast, and is too much of a
controversial system to allow anymore. Being at almost 50% for a vote is
bad enough, but seeing that WizDen chem is mostly fine? That's enough
for me.

:cl:
- remove: Removed FunkyChem. See PR for an explanation.

---------

Co-authored-by: TheDen-Bot <186991657+TheDen-Bot@users.noreply.github.com>