Re-Add some Funkychems (#984)
**Changelog**
<!-- Add a Changelog entry to make players aware of new features or
changes that could affect gameplay.
Make sure to read the guidelines and take this Changelog template out of
the comment block in order for it to show up.
Changelog must have a :cl: symbol, so the bot recognizes the changes and
adds them to the game's changelog. -->
<!--
:cl:
- add: Added fun!
- remove: Removed fun!
- tweak: Changed fun!
- fix: Fixed fun!
-->
:cl:
- add: Re-added salbutamol, oxandrolone, pentetic acid, granibitulari,
convermol, salicylic acid, deunomia
- add: Re-added liquid exotium, stasicillin, negacordrazine, stabilised
stimulants, and revivopine.

---------

Co-authored-by: TheDen-Bot <186991657+TheDen-Bot@users.noreply.github.com>