Revert "Keep Departmental Clothing Requirements to Any Department Members" & Other Loadout Tweaks/Additions (#955)
Reverts TheDenSS14/TheDen#932 and loosens a lot of restrictions on
clothing

Reasoning for Sec restrictions: Most of the items restricted are because
the variant they're the lower tier of is generally just an insignia on
the shoulder (silver = warden gold = hos)
or cause its the corpsman, who should be easily noticeable in a combat
scenario, as they're the combat medic

Reasoning for Robo restrictions: there are discussions of even further
defining the roboticist as a role, and opening all it's clothing to all
of epistemics devalues the role itself and borders on being, may as well
lose the role in it's entirety. Primarily post-woundmed.

Reasoning for basic bag restrictions: The roles that are restricted from
the basic bags have a variant of the given basic bag with equipment for
them (noted in the comment)

Other changes: 
Xor changed to Or to help make it more clear that those loadout items
work on (mostly) medical OR corpsman

:cl:
- add: Adds missing messenger bags to loadouts
- add: Boxer, Martial Artist, Zookeeper, Gladiator and Prison Guard are
added as loadout categories.
- add: Adds a lot of starting role items into loadouts.
- add: Gives all loadouts (that Adrian didn't touch) long coats if one
was available.
- tweak: Enables some items from specialist roles to be opened to the
entire department.
- tweak: Nukes the cost of all arm warmers/thongs/thigh highs/collars
except the valid ones (which remain at 1 cost)
- tweak: Changes a lot of loadout items, please recheck your loadouts
for any errors.
- remove: Blitz's sanity.

---------

Co-authored-by: TheDen-Bot <186991657+TheDen-Bot@users.noreply.github.com>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>