Fix Vent & Thermomachine Disassembly (#1106)
Fixes this [Discord
Thread](https://discord.com/channels/1301753657024319488/1394313641402175508)

## About the PR
Because of the deconstruction steps of the vents and thermomachines
conflicting with changing the pipe layer,
~~`PipeLayersLockedIfUnanchored` was added so vents can be
deconstructed~~ the interaction has been commented out. This was the
least egregious fix personally since alternatives are:
* removing the ability to weld shut a unanchored vent by it going
straight for sheets, or
 * using wirecutters instead of a screwdriver first.

It also includes a upstream cherry-pick fixing some edge cases with
popups.

**Changelog**
:cl:
- removed: Removed using a screwdriver directly on pipes to change its
layer as it conflicts with deconstruction of vents and thermomachines.
You can still right-click and use the verbs instead or hold the device
and use while having a screwdriver in the off-hand. This will be
revisited for a better solution.
- fixed: Vents and thermomachines can be deconstructed again.

---------

Co-authored-by: Perry Fraser <perryprog@users.noreply.github.com>
Co-authored-by: TheDen-Bot <186991657+TheDen-Bot@users.noreply.github.com>