Edge Security Medical Refurbishment (#1164)
<!-- Guidelines:
https://docs.spacestation14.io/en/getting-started/pr-guideline -->
<!-- NOTE: All code submitted to this repository is ALWAYS licensed
under the AGPL-3.0-or-later license.
The REUSE Specification headers or separate .license files indicate a
secondary license (e.g., MPL or MIT) solely to facilitate
integration for projects that do not use the AGPL license. This
secondary license does not replace the fact that AGPL-3.0-or-later
remains the primary and binding license.
Uncomment and modify the following line if you wish to change the
license from the default of AGPL.-->

<!--- LICENSE: AGPL -->
## Refurbished Security with an overhauled layout, updated Perma to be
open plan, made adjustments to medical for better use of space.

Complete overhaul of Security
Two new maints areas above Medical
Medical rooms moved around to allow for an additional chemist and a
larger therapists office.
Open plan Perma, aim to make it more comfortable and less claustrophobic
to residents

## Why / Balance
Access to brig via windows for crew to talk with prisoners.
Brigmedic next to perma for ease of treatment
Detectives office separated out to allow for more RP opportunities
Perma expanded to make it nicer
Wardens room between Armoury/Perma
Added in 'Office Sec' so that people could RP those elements if
interested
Added a date room above medical with views of space
Added in tider outlook for all our greytide friends
Moved Chemistry to a larger space to allow for two chemists on Edge
Moved Cryogenics closer to the front of Medical
Moved Paramedics room further out of the way
Swapped Cloning and Morgue to allow the Morgue to better utalise the
space
Expanded the size of the Therapists Office


## Media
When I work out how to export map picture

## Requirements
N/A


## Breaking changes
I hope not OwO

**Changelog**
:cl:
- add: Edge Station: Two additional maints rooms
- add: Edge Station:  Perma fishing
- tweak: Edge Station:  Medical Layout
- tweak: Edge Station: Security Layout 
- tweak: Edge Station:  Perma open plan

---------

Co-authored-by: TheDen-Bot <186991657+TheDen-Bot@users.noreply.github.com>