Chat QoL (#1435)
Fixed ' and , not removing the space between messages because of silly
goose moment.
Putting a ! at the front of the action formats it like (Name) Action
Putting a ' or a , at the front of the action removes the space between
Name and Action, only if it isn't starting with a !

<img width="305" height="172" alt="image"
src="https://github.com/user-attachments/assets/42cba9e0-5616-4f0e-84cb-c5c93f95a9ec"
/>

:cl:
- add: Added the ability to use a subtledetailed or medetailed format by
putting a ! at the front of your action in emote or subtle.
- fix: The space between name and message for subtle and emotes is now
properly removed when you put a ' or a , at the front.