Hellshift Removal (#1601)
Survival is now the hellshift, but nothings changed. It's still as
finnicky as ever.
Added the Roundstart Antags gamemode that lacks an event scheduler but
has roundstart 10 traitors max and 2 thieves. Admins can interfere as
they wish for a high danger.

:cl:
- add: Added the Roundstart Antags gamemode. This preset doesn't contain
an event scheduler, but has 10 traitors maximum which spawn based on 10%
of the population and 2 thieves. Admins are welcome to interfere as they
feel is warranted for a high danger shift.
- remove: Hellshift was removed from High Danger options.