Update Sprite Credits + Remove a Floof Plush Sprite (#2050)
as per ticket request; original spriter asked for an asset to be removed
and credited name updated.
floofstation already made the change of removing the sprites entirely;
we've verified the identity of the spriter.
this PR is not getting merged without the approval of the contributor

<img width="462" height="240" alt="image"
src="https://github.com/user-attachments/assets/7405d312-e8a8-4f5e-bb73-c3020e2be37d"
/>