Credit LandsharkRAWR for the Rose PR (#2121)
i meant to force push this to #2102 but i forgot to oops. because that
PR was lacking commit authorship data for the ported features, the SPDX
header script would miss the original authorship data for that
particular feature, and i am NOT standing for it. im sitting down

literally just comments